refactor(ai): Split chat panel into focused sub-components
Breaks the chat panel into the session sidebar, the message list,
the empty state, the composer and the result renderer as their own
components. The parent now only owns orchestration (data flow,
scrolling, focus handoff), which makes each piece easier to test and
restyle in isolation.
